Support overriding syncing behavior for multiple stores.
`sync_multiple` (and the inner `do_sync_multiple`) handle updating
persistent global state, which is necessary before we can sync
any engines. However, the Sync Manager provides a clients engine
which works differently than the others: it doesn't persist any state
except local client settings, it can't be disabled, declined, reset, or
wiped, its sync ID (and last sync time) isn't meaningful, it must sync
before any others, and failing to sync it must fail the entire sync.

This is also complicated by the fact that the `sync15` crate doesn't
know anything about the stores it's syncing—the Sync Manager handles
that. We could lift the clients engine into `sync15`, and provide a
delegate for handling commands...but that's awkward. We'd also like to
move away from the clients collection entirely in the future, and the
extra indirection we'd need to implement clients in `sync15` doesn't
seem worth it.

So, instead, we provide a `sync_multiple_with_params` function for
the Sync Manager to use. It'll take care of setting up global and
cached state. The `sync_all_stores` function is also exposed, which
we can call after we sync the clients engine.